switch to more robust mergedspikeanalysis
