+++ src/spikeanalysis/merged_spike_analysis.py
@@ -1,260 +1,122 @@
 from __future__ import annotations
-
-from dataclasses import dataclass
-from typing import Optional, Union, Literal
-
-
 import numpy as np
 
 from .spike_analysis import SpikeAnalysis
 from .curated_spike_analysis import CuratedSpikeAnalysis
 
 
-_merge_psth_values = ("zscore", "fr", "latencies", "isi", True)
+class MergedSpikeAnalysis(SpikeAnalysis):
 
+    def __init__(self, spikeanalysis_list=None, name_list=None, save_parameters=False, verbose=False):
 
-@dataclass
-class MergedSpikeAnalysis:
-    """class for merging neurons from separate animals for plotting"""
+        self.spikeanalysis_list = spikeanalysis_list or []
+        self.name_list = name_list or []
+        super().__init__(save_parameters=save_parameters, verbose=verbose)
 
-    spikeanalysis_list: list
-    name_list: list | None
+    def add_analysis(self, spikeanalysis, name):
 
-    def __post_init__(self):
-        if self.name_list is None:
-            pass
+        if isinstance(spikeanalysis, list):
+            if len(spikeanalysis) != len(name):
+                raise RuntimeError(f"{len(spikeanalysis)=} != {len(name)=}")
+            for idx, sa in enumerate(spikeanalysis):
+                self.spikeanalysis_list.append(sa)
+                self.name_list.append(name[idx])
         else:
-            assert len(self.spikeanalysis_list) == len(self.name_list), "each dataset needs a name value"
+            if not isinstance(spikeanalysis, (SpikeAnalysis | CuratedSpikeAnalysis)):
+                raise TypeError(f"Spikeanalysis must be a list or a spikeanalysis not a type {type(spikeanalysis)}")
+            if not isinstance(name, str):
+                raise TypeError("if spikeanalysis is type SpikeAnalysis, then name must be a string")
+            self.spikeanalysis_list.append(spikeanalysis)
+            self.name_list.append(name)
 
-    def add_analysis(
+    def merge_data(self):
+
+        self.events = self.spikeanalysis_list[0].events
+        self._sampling_rate = self.spikeanalysis_list[0]._sampling_rate
+        self.events_list = [sa.events for sa in self.spikeanalysis_list]
+
+        for stim in self.events.values():
+            for sa in self.spikeanalysis_list:
+                if self._get_stim_key(sa.events, stim["stim"]) is None:
+                    raise RuntimeError("Can not combine incompatible data")
+                if sa._sampling_rate != self._sampling_rate:
+                    raise RuntimeError("Can not combine incompatible data")
+
+        self._total_stim = len(self.events.keys())
+        spike_times = np.concatenate([sa.raw_spike_times for sa in self.spikeanalysis_list])
+
+        sub_cluster_ids = np.array(
+            [
+                f"{self.name_list[idx]}-{spike}"
+                for idx, sa in enumerate(self.spikeanalysis_list)
+                for spike in sa.spike_clusters
+            ]
+        )
+
+        sort_idx = np.argsort(spike_times)
+
+        self.raw_spike_times = spike_times[sort_idx]
+        self.spike_clusters = sub_cluster_ids[sort_idx]
+        self.cluster_ids = np.unique(self.spike_clusters)
+
+        event_names = {v["stim"]: k for k, v in self.events.items()}
+
+        self._matched_trial_groups = {}
+        for stim_name, stim_key in event_names.items():
+            tg_0 = set(self.events[stim_key]["trial_groups"])
+            tg_0_len = len(tg_0)
+            # start with easy case of checking if number of trial groups the same
+            if all(
+                [len(set(x[self._get_stim_key(x, stim_name)]["trial_groups"])) == tg_0_len for x in self.events_list]
+            ):
+                # if same number now we need to confirm same values
+                for event in self.events_list:
+                    for tg in tg_0:
+                        if tg not in event[self._get_stim_key(event, stim_name)]["trial_groups"]:
+                            unbalanced = True
+                            break
+                        else:
+                            unbalanced = False
+                    if unbalanced:
+                        self._matched_trial_groups[stim_name] = False
+                        break
+                    else:
+                        self._matched_trial_groups[stim_name] = True
+            else:
+                self._matched_trial_groups[stim_name] = False
+
+    def _get_stim_key(self, event, stim_name):
+
+        for key, value in event.items():
+            if value["stim"] == stim_name:
+                return key
+
+    def _fill_merged_data(
         self,
-        analysis: SpikeAnalysis | CuratedSpikeAnalysis | list[SpikeAnalysis | CuratedSpikeAnalysis],
-        name: str | None | list[str],
+        stim,
+        fill,
+        data_list,
     ):
-        """Function for adding an additional SpikeAnalysis to be merged
-        Parameters
-        ----------
-        analysis: SpikeAnalysis | CuratedSpikeAnalysis | list[SpikeAnalysis|CuratedSpikeAnalysis]
-            The analysis or analyses to add
-        name: str | None | list[str]
-             The names of the animal/dataset that is being added in analysis
-        """
-        if self.name_list is not None:
-            assert len(self.spikeanalysis_list) == len(self.name_list), "must provide name if other datasets named"
-            if isinstance(analysis, list):
-                assert isinstance(name, list), "if analysis is a list of analysis then name must be a list of names"
-                for idx, sa in enumerate(analysis):
-                    self.spikeanalysis_list.append(sa)
-                    self.name_list.append(name[idx])
-            else:
-                self.spikeanalysis_list.append(analysis)
-                self.name_list.append(name)
-        else:
-            print("other datasets were not given names ignoring naming")
-            if isinstance(analysis, list):
-                for sa in analysis:
-                    self.spikeanalysis_list.append(sa)
-            else:
-                self.spikeanalysis_list.append(analysis)
 
-    def merge(
-        self, psth: bool | list[Literal["zscore", "fr", "latencies", "isi"]] = True, stim_name: str | None = None
-    ):
-        """Function for merging different types of datasets between SpikeAnalysis objects
-        Parameters
-        ----------
-        psth: bool | list['zscore'|'fr'|'latencies'|'isi'], default: True
-            Indicates whether merging should occur at the raw psth level or other datasets
-        stim_name: str | None, default None
-            Whether to just analyze would stimulus subtype"""
-        # merge the cluster_ids for plotting
-        assert (
-            len(self.spikeanalysis_list) >= 2
-        ), f"merge should only be run on multiple datasets you currently have {len(self.spikeanalysis_list)} datasets"
+        if not np.isnan(fill) and not isinstance(fill, (int, float)):
+            raise TypeError(f"fill should be nan or ideally 0; it is {fill}")
 
-        assert isinstance(self.spikeanalysis_list[0].psths, dict), "must have psth to merge"
+        tg_list = [np.unique(x[self._get_stim_key(x, stim)]['trial_groups']) for x in self.events_list]
+        flat_tg_list = list(set([y for x in tg_list for y in x]))
+        for psth_idx, fr in enumerate(data_list):
+            current_tg = tg_list[psth_idx]
+            expand_dims = []
+            for sub_tg in flat_tg_list:
+                if sub_tg not in current_tg:
+                    expand_dims.append(flat_tg_list.index(sub_tg))
+            for dim in expand_dims:
+                fr = np.insert(fr, 1, fill, axis=dim)
+                data_list[psth_idx] = fr
 
-        if not isinstance(psth, bool):
-            for category in psth:
-                assert category in (
-                    "zscore",
-                    "fr",
-                    "latencies",
-                    "isi",
-                ), f"the only values you can use for psth are {_merge_psth_values}"
+    def get_raw_psth(self, window, time_bin_ms):
 
-        cluster_ids = []
-        for idx, sa in enumerate(self.spikeanalysis_list):
-            if isinstance(self.name_list, list):
-                sub_cluster_ids = [str(self.name_list[idx]) + str(cid) for cid in sa.cluster_ids]
-            else:
-                sub_cluster_ids = [str(idx) + str(cid) for cid in sa.cluster_ids]
-            cluster_ids.append(sub_cluster_ids)
-        final_cluster_ids = [cid for sub_cid in cluster_ids for cid in sub_cid]
-
-        self.cluster_ids = final_cluster_ids
-
-        # merge the events for plotting
-        events = {}
-        if stim_name is not None:
-            events[stim_name] = self.spikeanalysis_list[0].events[stim_name]
-        else:
-            events = self.spikeanalysis_list[0].events
-
-        self.events = events
-
-        if psth == True:
-            psths_list = []
-            for idx, sa in enumerate(self.spikeanalysis_list):
-                psths = sa.psths
-                merge_psth_dict = {}
-                psth_bins = {}
-                for sub_stim, psth_values in psths.items():
-                    merge_psth = psth_values["psth"]
-                    bins = psth_values["bins"]
-                    psth_bins[sub_stim] = bins
-                    merge_psth_dict[sub_stim] = merge_psth
-                    psths_list.append(merge_psth_dict)
-
-            data_merge = self._merge(psths_list, stim_name)
-
-            for key in data_merge.keys():
-                if key in psth_bins.keys():
-                    final_psth = data_merge[key]
-                    data_merge[key] = {}
-                    data_merge[key]["bins"] = psth_bins[key]
-                    data_merge[key]["psth"] = final_psth
-
-            self.data = data_merge
-            self.use_psth = True
-        else:
-            self.use_psth = False
-            z_list = []
-            fr_list = []
-            lat_list = []
-            isi_list = []
-            for idx, sa in enumerate(self.spikeanalysis_list):
-                if "zscore" in psth:
-                    z_list.append(sa.z_scores)
-                    z_bins = sa.z_bins
-                    z_windows = sa.z_windows
-                if "fr" in psth:
-                    fr_list.append(sa.mean_firing_rate)
-                    fr_bins = sa.fr_bins
-
-                if "latencies" in psth:
-                    raise NotImplementedError
-                if "isi" in psth:
-                    raise NotImplementedError
-
-            if len(z_list) != 0:
-                z_scores = self._merge(z_list, stim_name=stim_name)
-                self.z_scores = z_scores
-                self.z_bins = z_bins
-                self.z_windows = z_windows
-
-            if len(fr_list) != 0:
-                final_fr = self._merge(fr_list, stim_name=stim_name)
-                self.mean_firing_rate = final_fr
-                self.fr_bins = fr_bins
-
-            if len(lat_list) != 0:
-                raise NotImplementedError
-
-            if len(isi_list) != 0:
-                raise NotImplementedError
-
-    def _merge(self, dataset_list: list, stim_name: str) -> dict:
-        """Internal funtion for merging datasets"""
-        data_merge = {}
-        if stim_name is None:
-            for stim in dataset_list[0].keys():
-                data_merge[stim] = []
-                for dataset in dataset_list:
-                    data_merge[stim].append(dataset[stim])
-        else:
-            data_merge[stim_name] = []
-            for dataset in dataset_list:
-                data_merge[stim_name].append(dataset[stim_name])
-
-        for stim, array in data_merge.items():
-            data_merge[stim] = np.concatenate(array, axis=0)
-
-        return data_merge
-
-    def get_merged_data(self):
-        """function for returning an
-        instatiation of an MSA class for plotting
-
-        Returns
-        -------
-        msa: spikeanalysis.MSA
-            mergedspikeanalysis data for analysis"""
-        msa = MSA()
-        msa._set_cluster_ids(self.cluster_ids)
-        msa._set_events(self.events)
-        msa._sampling_rate = self.spikeanalysis_list[0]._sampling_rate
-
-        if self.use_psth:
-            msa.psths = self.data
-        else:
-            try:
-                msa.z_scores = self.z_scores
-                msa.z_bins = self.z_bins
-                msa.z_windows = self.z_windows
-            except AttributeError:
-                pass
-            try:
-                msa.mean_firing_rate = self.mean_firing_rate
-                msa.fr_bins = self.fr_bins
-            except AttributeError:
-                pass
-            try:
-                msa.latency = self.latency
-            except AttributeError:
-                pass
-            try:
-                msa.isi = self.isi
-                msa.isi_values = self.isi_values
-            except AttributeError:
-                pass
-
-        msa.use_psth = self.use_psth
-
-        return msa
-
-
-class MSA(SpikeAnalysis):
-    """Class for plotting and some analysis of merged data"""
-
-    def __init__(self):
-        self.use_psth = False
-        super().__init__()
-
-    def _set_cluster_ids(self, cluster_ids: np.array | list):
-        """used for setting the new cluster ids values"""
-        self.cluster_ids = cluster_ids
-
-    def _set_events(self, events: dict):
-        """method for setting new events"""
-        self.events = events
-
-    def get_raw_psth(self):
-        raise NotImplementedError
-
-    def set_spike_data(self):
-        print("data is immutable")
-
-    def set_stimulus_data(self):
-        print("data is immutable")
-
-    def z_score_data(
-        self, time_bin_ms: list[float] | float, bsl_window: list | list[list], z_window: list | list[list]
-    ):
-        if self.use_psth:
-            return super().z_score_data(time_bin_ms, bsl_window, z_window)
-        else:
-            raise NotImplementedError
+        for sa in self.spikeanalysis_list:
+            sa.get_raw_psth(window=window, time_bin_ms=time_bin_ms)
 
     def get_raw_firing_rate(
         self,
@@ -263,17 +125,59 @@ class MSA(SpikeAnalysis):
         mode: str,
         bsl_window: list | list[list] | None = None,
         sm_time_ms: list[float] | float | None = None,
+        fill=np.nan,
     ):
-        if self.use_psth:
-            return super().get_raw_firing_rate(time_bin_ms, fr_window, mode, bsl_window, sm_time_ms)
-        else:
-            raise NotImplementedError
+
+        firing_rates_list = []
+        for sa in self.spikeanalysis_list:
+            sa.get_raw_firing_rate(time_bin_ms, fr_window, mode, bsl_window, sm_time_ms)
+            firing_rates_list.append(sa.return_value("mean_firing_rate"))
+
+        merged_firing_rates = {}
+        for stim in firing_rates_list[0].keys():
+            prepped_firing_rates_list = [firing_rate[stim] for firing_rate in firing_rates_list]
+            if self._matched_trial_groups[stim]:
+                merged_firing_rates[stim] = np.concatenate(prepped_firing_rates_list, axis=0)
+            else:
+                self._fill_merged_data(stim, fill, prepped_firing_rates_list)
+                merged_firing_rates[stim] = np.concatenate(prepped_firing_rates_list, axis=0)
+
+        self.mean_firing_rate = merged_firing_rates
+        self.fr_bins = self.spikeanalysis_list[0].fr_bins
+
+    def z_score_data(self, time_bin_ms, bsl_window, z_window, eps=0, fill=np.nan):
+
+        z_score_list = []
+        for sa in self.spikeanalysis_list:
+            sa.z_score_data(time_bin_ms, bsl_window, z_window, eps)
+            z_score_list.append(sa.return_value("z_scores"))
+
+        merged_z_scores = {}
+        for stim in z_score_list[0].keys():
+            prepped_z_score_list = [z_score[stim] for z_score in z_score_list]
+            if self._matched_trial_groups[stim]:
+                merged_z_scores[stim] = np.concatenate(prepped_z_score_list, axis=0)
+            else:
+                self._fill_merged_data(stim, fill, prepped_z_score_list)
+                merged_z_scores[stim] = np.concatenate(prepped_z_score_list, axis=0)
+
+        self.z_scores = merged_z_scores
+        self.z_bins = self.spikeanalysis_list[0].z_bins
+
+    def latencies(self):
+        print("To do")
 
     def get_interspike_intervals(self):
-        raise NotImplementedError
+        super().get_interspike_intervals()
 
-    def compute_event_interspike_intervals(self):
-        raise NotImplementedError
+    def compute_event_interspike_intervals(self, time_ms: float = 200):
+        raise NotImplementedError("Should run in the base Spikeanalysis for this")
 
-    def autocorrelogram(self):
-        raise NotImplementedError
+    def trial_correlation(
+        self,
+        window: list | list[list],
+        time_bin_ms: float | None = None,
+        dataset = "psth",
+        method = "pearson",
+    ):
+        raise NotImplementedError("Should run in the base SpikeAnalysis")
