+++ src/spikeanalysis/spike_analysis.py
@@ -820,7 +820,7 @@ class SpikeAnalysis:
         self.acg = acg
 
     def return_value(self, value: str):
-        _values = ("zscores", "raw_zscores", "mean_firing_rate", "raw_firing_rate", "correlations", "latency", "psths")
+        _values = ("z_scores", "raw_zscores", "mean_firing_rate", "raw_firing_rate", "correlations", "latency", "psths")
 
         if hasattr(self, value):
             return getattr(self, value)
