+++ test/test_merged_spike_analysis.py
@@ -3,7 +3,7 @@ import numpy as np
 from pathlib import Path
 
 
-from spikeanalysis.merged_spike_analysis import MergedSpikeAnalysis, MSA
+from spikeanalysis.merged_spike_analysis import MergedSpikeAnalysis
 from spikeanalysis.stimulus_data import StimulusData
 from spikeanalysis.spike_data import SpikeData
 from spikeanalysis.spike_analysis import SpikeAnalysis
@@ -24,104 +24,23 @@ def sa():
     return spiketrain
 
 
-def test_init(sa):
-    test_msa = MergedSpikeAnalysis([sa, sa], name_list=None)
+def test_ma_init(sa):
+    test_msa = MergedSpikeAnalysis()
 
-    assert isinstance(test_msa, MergedSpikeAnalysis)
-
-
-def test_init_names(sa):
-    test_msa = MergedSpikeAnalysis(spikeanalysis_list=[sa, sa], name_list=["test", "test1"])
-
-    assert isinstance(test_msa, MergedSpikeAnalysis)
-
-
-def test_init_failure(sa):
-    with pytest.raises(AssertionError):
-        test_msa = MergedSpikeAnalysis(
-            spikeanalysis_list=[sa, sa],
-            name_list=[
-                "test",
-            ],
-        )
+    assert len(test_msa.spikeanalysis_list) == 0, "list should be empty"
+    assert not test_msa._save_params
 
 
 def test_add_analysis(sa):
-    test_msa = MergedSpikeAnalysis(spikeanalysis_list=[sa, sa], name_list=["test", "test1"])
-    test_msa.add_analysis(sa, "test2")
 
-    assert len(test_msa.spikeanalysis_list) == 3
+    test_msa = MergedSpikeAnalysis()
+    test_msa.add_analysis([sa, sa], name=["test1", "test2"])
 
-    test_msa.add_analysis([sa, sa], ["test3", "test4"])
-
-    assert len(test_msa.spikeanalysis_list) == 5
-    assert "test4" in test_msa.name_list
-
-    test_msa_no_name = MergedSpikeAnalysis([sa, sa], name_list=None)
-    test_msa_no_name.add_analysis(sa, name=None)
-
-    assert len(test_msa_no_name.spikeanalysis_list) == 3
-    test_msa_no_name.add_analysis([sa, sa], name=None)
-    assert len(test_msa_no_name.spikeanalysis_list) == 5
+    assert len(test_msa.spikeanalysis_list) == len(test_msa.name_list)
 
 
-def test_merge_psth(sa):
-    sa.events = {
-        "0": {"events": np.array([100]), "lengths": np.array([200]), "trial_groups": np.array([1]), "stim": "test"}
-    }
-    sa.get_raw_psth(
-        window=[0, 300],
-        time_bin_ms=50,
-    )
-
-    test_msa = MergedSpikeAnalysis([sa, sa], name_list=["test", "test1"])
-
-    test_msa.merge(psth=True)
-    test_merged_msa = test_msa.get_merged_data()
-
-    assert isinstance(test_merged_msa.cluster_ids, list)
-    print(test_merged_msa.cluster_ids)
-    assert len(test_merged_msa.cluster_ids) == 4
-
-    assert isinstance(test_merged_msa.events, dict)
-
-    psth = test_merged_msa.psths["test"]["psth"]
-    assert np.shape(psth) == (4, 1, 6000)
-
-    assert isinstance(test_merged_msa, SpikeAnalysis)
-    assert isinstance(test_merged_msa, MSA)
-
-    with pytest.raises(NotImplementedError):
-        test_merged_msa.get_raw_psth()
-    with pytest.raises(NotImplementedError):
-        test_merged_msa.get_interspike_intervals()
-    with pytest.raises(NotImplementedError):
-        test_merged_msa.autocorrelogram()
-
-
-def test_merge_z_score(sa):
-    sa.events = {
-        "0": {"events": np.array([100]), "lengths": np.array([200]), "trial_groups": np.array([1]), "stim": "test"}
-    }
-    sa.get_raw_psth(
-        window=[0, 300],
-        time_bin_ms=50,
-    )
-    sa.z_score_data(time_bin_ms=1000, bsl_window=[0, 50], z_window=[0, 300])
-
-    test_msa = MergedSpikeAnalysis([sa, sa], name_list=["test", "test1"])
-
-    with pytest.raises(AssertionError):
-        test_msa.merge(psth=["zscoresa"])
-
-    test_msa.merge(psth=["zscore"])
-    test_merged_msa = test_msa.get_merged_data()
-
-    assert isinstance(test_merged_msa.z_scores, dict)
-
-    test_merged_msa.set_stimulus_data()
-    test_merged_msa.set_spike_data()
-
+@pytest.fixture(scope="module")
+def sa_mocked(sa):
     sa.events = {
         "0": {
             "events": np.array([100, 200]),
@@ -130,12 +49,76 @@ def test_merge_z_score(sa):
             "stim": "test",
         }
     }
-    sa.get_raw_psth(window=[0, 300], time_bin_ms=50)
-    sa.get_raw_firing_rate(time_bin_ms=1000, bsl_window=None, fr_window=[0, 300], mode="raw")
-    sa.z_score_data(time_bin_ms=1000, bsl_window=[0, 50], z_window=[0, 300])
 
-    test_msa = MergedSpikeAnalysis([sa, sa], name_list=["test", "test1"])
-    test_msa.merge(psth=["zscore", "fr"])
-    test_merged_msa = test_msa.get_merged_data()
+    return sa
 
-    assert isinstance(test_merged_msa.mean_firing_rate, dict)
+
+def test_merge(sa_mocked):
+
+    test_msa = MergedSpikeAnalysis([sa_mocked, sa_mocked], ["test1", "test2"])
+    test_msa.merge_data()
+
+    assert len(test_msa.raw_spike_times) == 2 * len(sa_mocked.raw_spike_times)  # same data twice
+    assert "test1-1" in test_msa.cluster_ids
+    assert test_msa._matched_trial_groups["test"]
+
+
+def test_fr_z_psth(sa_mocked):
+
+    test_msa = MergedSpikeAnalysis([sa_mocked, sa_mocked], ["test1", "test2"])
+    test_msa.merge_data()
+    test_msa.get_raw_psth(
+        window=[0, 300],
+        time_bin_ms=50,
+    )
+
+    test_msa.get_raw_firing_rate(time_bin_ms=1000, bsl_window=None, fr_window=[0, 300], mode="raw")
+
+    assert test_msa.mean_firing_rate["test"].shape[0] == 4  # sa_mocked has two neurons
+    assert test_msa.mean_firing_rate["test"].shape[1:] == sa_mocked.mean_firing_rate["test"].shape[1:]
+
+    test_msa.z_score_data(time_bin_ms=1000, bsl_window=[0, 50], z_window=[0, 300])
+
+    assert test_msa.z_scores["test"].shape[0] == 4
+
+    assert test_msa.z_scores["test"].shape[1:] == sa_mocked.z_scores["test"].shape[1:]  # same except neuron number
+
+
+def test_fr_z_psth_different_trials(sa_mocked):
+    from copy import deepcopy
+
+    sa_mocked1 = deepcopy(sa_mocked)
+    sa_mocked1.events = {
+        "0": {
+            "events": np.array([100, 150, 200, 250]),
+            "lengths": np.array([100, 100, 100, 100]),
+            "trial_groups": np.array([1, 2, 1, 2]),
+            "stim": "test",
+        }
+    }
+
+    test_msa = MergedSpikeAnalysis([sa_mocked, sa_mocked1], ["test1", "test2"])
+    test_msa.merge_data()
+    test_msa.get_raw_psth(
+        window=[0, 300],
+        time_bin_ms=50,
+    )
+    test_msa.get_raw_firing_rate(time_bin_ms=1000, bsl_window=None, fr_window=[0, 300], mode="raw")
+
+    assert test_msa.mean_firing_rate["test"].shape[0] == 4  # sa_mocked has two neurons
+    
+
+    assert np.isnan(test_msa.mean_firing_rate["test"][0, 1, 0])  # should fill with nan
+
+    test_msa.z_score_data(time_bin_ms=1000, bsl_window=[0, 50], z_window=[0, 300])
+
+    assert test_msa.z_scores["test"].shape[0] == 4
+
+
+def test_interspike_interval(sa_mocked):
+
+    test_msa = MergedSpikeAnalysis([sa_mocked, sa_mocked], ["test1", "test2"])
+    test_msa.merge_data()
+    test_msa.get_interspike_intervals()
+    
+    assert isinstance(test_msa.isi_raw, dict)
